checkpoint 2

+++ backend/database/migrations/2025_08_09_100428_create_user_subscriptions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->uuid('uuid')->unique();
 
-            $table->integer('amount')->default(0);
+            $table->unsignedBigInteger('amount')->default(0);
 
             $table->boolean('is_paid')->default(false);
             $table->timestamp('expire_at')->nullable();
